+++ aasm/parsing/op/deftype.py
@@ -21,6 +21,10 @@ def op_DEFTYPE(
 ) -> None:
     state.require(state.agent_exists(agent_name), f"Agent {agent_name} is not defined.")
     state.require(
+        not state.last_graph.is_agent_defined(agent_name),
+        f"Agent {agent_name} is already defined.",
+    )
+    state.require(
         state.in_graph,
         "Cannot define agent type outside of graph scope.",
         "Try defining new graph using GRAPH.",